Refactor canvas
