+++ src/components/Customizing/index.scss
@@ -1,0 +1,30 @@
+.customizing {
+  &-title {
+    width: 300px;
+    height: 60px;
+    
+    padding-left: 1rem;
+    padding-bottom: 1rem;
+
+    display: flex;
+    align-items: flex-end;
+
+    color: #c4c4c4;
+  }
+
+  &-board {
+    width: 300px;
+    height: 70vh;
+  
+    margin: 0 30px;
+    padding: 20px;
+    border-radius: 20px;
+  
+    background: #ffffff;
+
+    input {
+      margin-top: 10px;
+      margin-bottom: 15px;
+    }
+  }
+}
