+++ src/components/Customizing/index.tsx
@@ -1,0 +1,10 @@
+function Customizing() {
+  return (
+    <>
+      <div className='customizing-title' />
+      <div className='customizing-board' />
+    </>
+  );
+}
+
+export default Customizing;
