+++ src/styles/Canvas.scss
@@ -13,65 +13,8 @@
   flex-direction: column;
 }
 
-.customizing {
-  &-title {
-    width: 300px;
-    height: 60px;
-    
-    padding-left: 1rem;
-    padding-bottom: 1rem;
-
-    display: flex;
-    align-items: flex-end;
-
-    color: #c4c4c4;
-  }
-
-  &-board {
-    width: 300px;
-    height: 70vh;
-  
-    margin: 0 30px;
-    padding: 20px;
-    border-radius: 20px;
-  
-    background: #ffffff;
-
-    input {
-      margin-top: 10px;
-    }
-  }
-}
-
-
 .adding {
   flex: 1;
-
-  &-board {
-    display: flex;
-    align-items: center;
-  
-    width: 500px;
-    height: 60px;
-    padding: 10px;
-    border-radius: 20px;
-
-    & >button {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    
-      width: 40px;
-      height: 40px;
-    
-      margin: 0 5px;
-      border: 1px solid #ffffff30;
-      border-radius: 50%;
-      background: none;
-      cursor: pointer;
-      color: #c4c4c4;
-    }
-  }
 }
 
 .canvas {
